requirements.txt
python-chess==0.23.9
slackclient==1.2.1
redis==2.10.6
prettytable==0.7.2
requests==2.19.1